+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.3.5
+
+- To address the issue of crashes occurring when using the NoticeFailure method on background threads, we have added StartTime and EndTime parameters to the method. This enhancement should prevent such crashes from happening.
+- Upgraded native iOS Agent to 7.4.11
+- Upgraded native Android Agent to 7.3.0
+
 ## 1.3.4
 
 - Upgraded native iOS Agent to 7.4.10